+++ packages/test/src/languages/library.ts
@@ -1,8 +1,13 @@
-import {builtinPrimitives, LanguageFactory} from "@lionweb/core"
-import {hashingIdGen} from "@lionweb/utilities"
+import {builtinPrimitives, chain, concatenator, LanguageFactory, lastOf} from "@lionweb/core"
+import {hasher} from "@lionweb/utilities"
 
 
-const factory = new LanguageFactory("library", "1", hashingIdGen())
+const factory = new LanguageFactory(
+    "library",
+    "1",
+    chain(concatenator("-"), hasher()),
+    lastOf
+)
 export const libraryLanguage = factory.language
 
 const {integerDatatype, stringDatatype} = builtinPrimitives