+++ packages/test/src/languages/multi.ts
@@ -1,9 +1,14 @@
-import {LanguageFactory} from "@lionweb/core"
-import {hashingIdGen} from "@lionweb/utilities"
+import {chain, concatenator, LanguageFactory, lastOf} from "@lionweb/core"
+import {hasher} from "@lionweb/utilities"
 import {library, libraryLanguage} from "./library.js"
 
 
-const factory = new LanguageFactory("multi", "1", hashingIdGen())
+const factory = new LanguageFactory(
+    "multi",
+    "1",
+    chain(concatenator("-"), hasher()),
+    lastOf
+)
 export const multiLanguage = factory.language.dependingOn(libraryLanguage)
 
 const container = factory.concept("Container", false)