+++ packages/test/src/m3/constraints.test.ts
@@ -1,38 +1,38 @@
 import {assert} from "chai"
-const {deepEqual} = assert
-
-import {Concept, issuesLanguage, Language, LanguageFactory} from "@lionweb/core"
+import {chain, concatenator, Concept, issuesLanguage, Language, LanguageFactory, lastOf} from "@lionweb/core"
 import {nanoIdGen} from "@lionweb/utilities"
+
+const {deepEqual} = assert
 
 
 describe("constraints (LionCore)", () => {
 
     it("check that inheritance cycles are detected", () => {
-        const factory = new LanguageFactory("metamodel", "1", nanoIdGen())
+        const factory = new LanguageFactory("metamodel", "1", chain(concatenator("-"), nanoIdGen()), lastOf)
         const {language} = factory
-        const cis = [0, 1, 2].map((i) => factory.interface(`interface ${i}`))
-        cis[2].extends.push(cis[1])
-        cis[1].extends.push(cis[0])
-        cis[0].extends.push(cis[2])
-        language.entities.push(...cis)
+        const intfaces = [0, 1, 2].map((i) => factory.interface(`interface ${i}`))
+        intfaces[2].extends.push(intfaces[1])
+        intfaces[1].extends.push(intfaces[0])
+        intfaces[0].extends.push(intfaces[2])
+        language.entities.push(...intfaces)
 
         const issues = issuesLanguage(language)
         deepEqual(issues.length, 3)
-        const message1 = issues?.find(({location}) => location === cis[0])?.message
+        const message1 = issues?.find(({location}) => location === intfaces[0])?.message
         deepEqual(message1, `A Interface can't inherit (directly or indirectly) from itself, but metamodel.interface 0 does so through the following cycle: metamodel.interface 0 -> metamodel.interface 2 -> metamodel.interface 1 -> metamodel.interface 0`)
     })
 
     it("check that trivial inheritance cycles are detected", () => {
-        const factory = new LanguageFactory("metamodel", "1", nanoIdGen())
+        const factory = new LanguageFactory("metamodel", "1", chain(concatenator("-"), nanoIdGen()), lastOf)
         const {language} = factory
-        const ci = factory.interface(`foo`)
-        ci.extends.push(ci)
-        language.entities.push(ci)
+        const intface = factory.interface(`foo`)
+        intface.extends.push(intface)
+        language.entities.push(intface)
 
         const issues = issuesLanguage(language)
         deepEqual(issues.length, 1)
         const {location, message} = issues[0]
-        deepEqual(location, ci)
+        deepEqual(location, intface)
         deepEqual(message, `A Interface can't inherit (directly or indirectly) from itself, but metamodel.foo does so through the following cycle: metamodel.foo -> metamodel.foo`)
     })