+++ packages/test/src/m3/key-generation.test.ts
@@ -1,14 +1,19 @@
 import {assert} from "chai"
 const {deepEqual} = assert
 
-import {builtinPrimitives, LanguageFactory, qualifiedNameBasedKeyGenerator} from "@lionweb/core"
-import {hashingIdGen} from "@lionweb/utilities"
+import {builtinPrimitives, chain, concatenator, LanguageFactory} from "@lionweb/core"
+import {hasher} from "@lionweb/utilities"
 
 
 describe("key generation", () => {
 
     it("based on qualified name", () => {
-        const factory = new LanguageFactory("FormLanguage", "1", hashingIdGen(), qualifiedNameBasedKeyGenerator("-"))
+        const factory = new LanguageFactory(
+            "FormLanguage",
+            "1",
+            chain(concatenator("-"), hasher()),
+            concatenator("-")
+        )
 
         const form = factory.concept("Form", false)
         factory.language.havingEntities(form)