+++ packages/utilities/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.5.0
 
+* _Breaking change:_ Change ID generation to hashing (`hasher` function).
+* (Make small fix in `asText` function.)
 * Add a function `tsTypesForLanguage` that generates TypeScript types.
 * Receive `asText` and ID checking from `@lionweb/core`.